Refactor code and enhance bot capabilities in discord server
This commit encompasses the deletion and modification of certain files to better organize the discord server functionalities. Major changes include the removal of the 'play' music command, addition of the 'ticket_create' modal and button feature. The code has also been refactored to help manage bot responses according to the interaction type (commands, buttons, modals), thereby increasing the interaction modalities of the bot.